feat: move LanguageClientImpl to a package which is exported without restrictions and rename it
