+++ org.eclipse.lsp4e/schema/languageServer.exsd
@@ -79,10 +79,10 @@
          <attribute name="clientImpl" type="string">
             <annotation>
                <documentation>
-                  An optional language client implementation to use as local endpoint. If undefined, the default implementation &lt;code&gt;org.eclipse.lsp4e.LanguageClientImpl&lt;/code&gt; is used. See the &lt;a href=&quot;https://github.com/eclipse-lsp4j/lsp4j/tree/main/documentation&quot;&gt;LSP4J documentation&lt;/a&gt; for more information on how to extend the Language Server Protocol.
+                  An optional language client implementation to use as local endpoint. If undefined, the default implementation &lt;code&gt;org.eclipse.lsp4e.client.DefaultLanguageClient&lt;/code&gt; is used. See the &lt;a href=&quot;https://github.com/eclipse-lsp4j/lsp4j/tree/main/documentation&quot;&gt;LSP4J documentation&lt;/a&gt; for more information on how to extend the Language Server Protocol.
                </documentation>
                <appinfo>
-                  <meta.attribute kind="java" basedOn="org.eclipse.lsp4e.LanguageClientImpl:"/>
+                  <meta.attribute kind="java" basedOn="org.eclipse.lsp4e.client.DefaultLanguageClient:"/>
                </appinfo>
             </annotation>
          </attribute>
